Refactor: extract menu open state decision to state package
